feat(#30) :: 레이아웃 처리
