+++ src/app/write/layout.tsx
@@ -1,22 +1,4 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
 import { Header } from "@/components/layout/Header";
-
-const geistSans = Geist({
-    variable: "--font-geist-sans",
-    subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-    variable: "--font-geist-mono",
-    subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-    title: "Create Next App",
-    description: "Generated by create next app",
-};
 
 export default function RootLayout({
                                        children,
@@ -24,13 +6,9 @@ export default function RootLayout({
     children: React.ReactNode;
 }>) {
     return (
-        <html lang="ko">
-        <body
-            className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-        >
-        <Header />
-        {children}
-        </body>
-        </html>
+      <div>
+          <Header />
+          {children}
+      </div>
     );
 }
